Add pselect6 and ppoll cases to the select test.
Not sure about this. It was failing even before I started, seemingly
because it doesn't like the debug symbols on the libc in my
cross-compiler setup, so kind of hard to test whether I've got it
right. It at least doesn't look like it's failing any more now than it
was before.